bin/test: Set up app sec group for tests
In some tests, we expect to be able to reach the external application
routes (i.e. things in the apps domain).  On deployments such as
minikube, that will be an internal IP (e.g. 10.x.x.x), and therefore
would not be reachable from applications unless we explicitly whitelist
them via additional application security groups.

This sets up a temporary security group for the specific IP (assuming
the apps domain resolves to a single IP, rather than some sort of set up
with multiple IP addresses), in the hopes that the tests would pass.